fix: handle change behaviour in newer kernels when POLLHUP is received
bfafe5ef changed when the kernel is creating the POLLHUP event on the seccomp notify fd.
It moved the call to seccomp_filter_release from release_task to do_exit, thus the event is now already triggered when the last task exits, not only once it has been waited for.

Therefore, we can no longer assume that a prior waitid call must have already collected the return code when we see this event.
Instead we check if we are still waiting for a result when the event is received and collect the return code if required.

fixes #1